moved EOExecutor report making to visualization package
