+++ core/requirements.txt
@@ -1,9 +1,6 @@
-numpy>=1.16
 attrs>=18.2.0
 python-dateutil
-matplotlib
-jinja2
-pygments
-sentinelhub>=2.5.0
+numpy>=1.16
 geopandas
+sentinelhub>=2.5.0
 tqdm>=4.27
