+++ core/setup.py
@@ -26,18 +26,19 @@ def get_version():
     return version
 
 
-setup(name='eo-learn-core',
-      python_requires='>=3.5',
-      version=get_version(),
-      description='Core Machine Learning Framework at Sinergise',
-      long_description=get_long_description(),
-      long_description_content_type='text/markdown',
-      url='https://github.com/sentinel-hub/eo-learn',
-      author='Sinergise EO research team',
-      author_email='eoresearch@sinergise.com',
-      license='MIT',
-      packages=find_packages(),
-      package_data={'eolearn': ['core/report_templates/report.html']},
-      include_package_data=True,
-      install_requires=parse_requirements("requirements.txt"),
-      zip_safe=False)
+setup(
+    name='eo-learn-core',
+    python_requires='>=3.5',
+    version=get_version(),
+    description='Core Machine Learning Framework at Sinergise',
+    long_description=get_long_description(),
+    long_description_content_type='text/markdown',
+    url='https://github.com/sentinel-hub/eo-learn',
+    author='Sinergise EO research team',
+    author_email='eoresearch@sinergise.com',
+    license='MIT',
+    packages=find_packages(),
+    include_package_data=True,
+    install_requires=parse_requirements("requirements.txt"),
+    zip_safe=False
+)
