+++ visualization/eolearn/visualization/__init__.py
@@ -4,6 +4,7 @@ A collection of visualization utilities
 
 from .eopatch_visualization import EOPatchVisualization
 from .eoworkflow_visualization import EOWorkflowVisualization
+from .eoexecutor_visualization import EOExecutorVisualization
 
 
 __version__ = '0.5.0'
