+++ visualization/requirements.txt
@@ -5,3 +5,6 @@ pyepsg
 xarray
 pydot
 graphviz>=0.10.1
+matplotlib
+jinja2
+pygments
